Tests/LibWeb/Text/expected/css/invalidate-sibling-affected-by-has.txt

background-color before: rgb(204, 204, 255)
background-color after: rgb(238, 238, 255)
